server/requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-multipart
numpy
thefuzz
python-Levenshtein
scikit-learn
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy[asyncio]
aiosqlite
asyncpg
python-dotenv
google-generativeai
pyarrow
pydantic-settings
python-json-logger
openai
